split tablet channel content
